VIH-7978 User Api Client Wrapper
