+++ BookingsApi/BookingsApi/BookingsApi.csproj
@@ -42,7 +42,6 @@
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="UserApi.Client" Version="1.24.3" />
     <PackageReference Include="VH.Core.Configuration" Version="0.1.12" />
     <PackageReference Include="ZymLabs.NSwag.FluentValidation" Version="0.4.0" />
   </ItemGroup>
